+++ Tests/TuistKitTests/Linter/TargetLinterTests.swift
@@ -14,6 +14,21 @@ final class TargetLinterTests: XCTestCase {
         subject = TargetLinter(fileHandler: fileHandler)
     }
 
+    func test_lint_when_target_has_invalid_bundle_identifier() {
+        let XCTAssertInvalidBundleId: (String) -> Void = { bundleId in
+            let bundleId = "_.company.app"
+            let target = Target.test(bundleId: bundleId)
+            let got = self.subject.lint(target: target)
+            let reason = "Invalid bundle identifier '\(bundleId)'. This string must be a uniform type identifier (UTI) that contains only alphanumeric (A-Z,a-z,0-9), hyphen (-), and period (.) characters."
+            XCTAssertTrue(got.contains(LintingIssue(reason: reason, severity: .error)))
+        }
+
+        XCTAssertInvalidBundleId("_.company.app")
+        XCTAssertInvalidBundleId("com.company.◌́")
+        XCTAssertInvalidBundleId("Ⅻ.company.app")
+        XCTAssertInvalidBundleId("ؼ.company.app")
+    }
+
     func test_lint_when_target_no_source_files() {
         let target = Target.test(sources: [])
         let got = subject.lint(target: target)